fix circular imports that are affecting Python 3.6 but not 3.7
